print last digit

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -24,6 +24,6 @@ int print_last_digit(int n)
 	}
 
 	last_digit = n % 10;
-	_putchar((int)last_digit);
+	_putchar(last_digit);
 	return (last_digit);
 }
